refactor macros into more files and start handling generics
